Fix type hint of hebbian learning rate arg

+++ HopfieldNetworkPyTorch/HopfieldNetwork/LearningRule/hebbian.py
@@ -3,7 +3,7 @@ from . import AbstractLearningRule
 import torch
 
 class HebbianLearningRule(AbstractLearningRule):
-    def __init__(self, learningRate: int = 1, maxEpochs: int = 1):
+    def __init__(self, learningRate: float = 1, maxEpochs: int = 1):
         super().__init__(learningRate, maxEpochs)
 
     def __call__(self, states: torch.Tensor) -> torch.Tensor:
